pyproject.toml
[project]
name = "claude-skills"
version = "0.1.0"
description = "Claude Code skills collection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.2.0",
]

[tool.uv]
package = false
